almost finished reafctoring
 for making event_loop, widget and window be independent. There're still some problems with working outside but at least inside works as expected.
