+++ reapy/core/gui/events.py
@@ -1,13 +1,37 @@
 import typing as ty
 import typing_extensions as te
 
+import atexit
+from enum import Enum, auto
+
 import reapy
 from reapy.core import ReapyObject
 from .singleton import Singleton, UUID, SingletonMeta
+from . import JS
+
+
+class EventTarget(Enum):
+    """Used to specify event targets relatively for EventClient."""
+    everyone = auto()
+    parent = auto()
+    childs = auto()
+    self_ = auto()
 
 
 class Event(ReapyObject):
     pass
+
+
+class EvFrame(Event):
+    """Fired every loop frame."""
+
+
+class EvStart(Event):
+    """Fired on mainloop start."""
+
+
+class EvExit(Event):
+    """Fired on mainloop exit."""
 
 
 class EventQueueItem(te.TypedDict):
@@ -27,25 +51,7 @@ class EventClientMeta(SingletonMeta):
         return obj  # type:ignore
 
 
-class EventClient(Singleton, metaclass=EventClientMeta):
-    event_handler: 'EventHandler'
-
-    def __init__(self, event_handler: 'EventHandler') -> None:
-        self.event_handler = event_handler
-
-    def on_event(self, event: Event) -> None:
-        """Handle general event.
-
-        Note
-        ----
-        Usually, this method should filter events interesting to the
-        particular class, then passthrough other events with or without
-        filtered event to the super().on_event().
-
-        Parameters
-        ----------
-        event : Event
-        """
+ListOfEventOrUUID = ty.Union[ty.List['EventClient'], ty.List[UUID]]
 
 
 class EventHandler(Singleton):
@@ -58,7 +64,7 @@ class EventHandler(Singleton):
     * create Handler and assign to event loop provider (currently — TopLevel)
     * ensure provider has the same EventHandler as from outside as from inside
     * fire events from the inside or outside loop
-    * fire queue every otside frame (deferred call)
+    * fire queue every outside frame (deferred call)
     * PROFIT!
 
     Note
@@ -110,17 +116,24 @@ class EventHandler(Singleton):
     def clear_queue(self) -> None:
         self._queue.clear()
 
-    def fire_event(self, event: Event, clients: ty.List[EventClient]) -> None:
+    def fire_event(
+        self, event: Event, clients: ty.Optional[ListOfEventOrUUID]
+    ) -> None:
         """Put event to queue or launch corresponding callbacks on clients.
 
         Parameters
         ----------
         event : Event
-        clients : List[EventClient]
+        clients : Optional[List[EventClient]]
+            If None — fires to every registered client.
         """
+        if clients is None:
+            clients = list(self.clients.values())
         if self.buffered and reapy.is_inside_reaper():
             return self.event_to_queue(event, clients)
         for client in clients:
+            if not isinstance(client, EventClient):
+                client = self.clients[client]
             client.on_event(event)
 
     def fire_queue(self) -> None:
@@ -133,9 +146,7 @@ class EventHandler(Singleton):
             )
         self.clear_queue()
 
-    def event_to_queue(
-        self, event: Event, clients: ty.List[EventClient]
-    ) -> None:
+    def event_to_queue(self, event: Event, clients: ListOfEventOrUUID) -> None:
         """Put event to queue for being fired from outside.
 
         Note
@@ -149,10 +160,20 @@ class EventHandler(Singleton):
         clients : List[EventClient]
         """
         self._queue.append(
-            EventQueueItem(event=event, clients=[cl.uuid for cl in clients])
+            EventQueueItem(
+                event=event,
+                clients=[
+                    cl.uuid if isinstance(cl, EventClient) else cl
+                    for cl in clients
+                ]
+            )
         )
 
-    def register_client(self, client: EventClient) -> None:
+    @reapy.inside_reaper()
+    def _register_client_inside(self, client: 'EventClient') -> None:
+        self.register_client(client)
+
+    def register_client(self, client: 'EventClient') -> None:
         """Let EventHandler recognize client later.
 
         Parameters
@@ -160,3 +181,225 @@ class EventHandler(Singleton):
         client : EventClient
         """
         self.clients[client.uuid] = client
+        if not reapy.is_inside_reaper():
+            self._register_client_inside(client)
+
+    @reapy.inside_reaper()
+    def _remove_client_inside(self, client: 'EventClient') -> None:
+        self.remove_client(client)
+
+    def remove_client(self, client: 'EventClient') -> None:
+        """Let EventHandler recognize client later.
+
+        Parameters
+        ----------
+        client : EventClient
+        """
+        del self.clients[client.uuid]
+        if not reapy.is_inside_reaper():
+            self._remove_client_inside(client)
+
+
+class EventClient(Singleton, metaclass=EventClientMeta):
+    event_handler: 'EventHandler'
+    parent: ty.Optional['EventClient']
+    childs: ty.Dict[UUID, 'EventClient']
+
+    def __init__(self, event_handler: 'EventHandler') -> None:
+        self.event_handler = event_handler
+        self.parent = None
+        self.childs = {}
+
+    @property
+    def _state(self) -> ty.Dict[str, object]:
+        return {
+            **super()._state, 'parent': self.parent,
+            'childs': self.childs,
+            'event_handler': self.event_handler
+        }
+
+    def on_event(self, event: Event) -> None:
+        """Handle general event.
+
+        Note
+        ----
+        Usually, this method should filter events interesting to the
+        particular class, then pass through other events with or without
+        filtered event to the super().on_event().
+
+        Parameters
+        ----------
+        event : Event
+        """
+        if isinstance(event, EvFrame):
+            return self.run()
+        if isinstance(event, EvExit):
+            return self.cleanup()
+        if isinstance(event, EvStart):
+            return self.setup()
+        return None
+
+    def fire_event(
+        self,
+        event: Event,
+        clients: ty.Union[ty.List['EventClient'],
+                          EventTarget] = EventTarget.childs
+    ) -> None:
+        if isinstance(clients, ty.Iterable):
+            return self.event_handler.fire_event(event, clients)
+        if clients == EventTarget.everyone:
+            return self.event_handler.fire_event(event, None)
+        if clients == EventTarget.childs:
+            return self.event_handler.fire_event(
+                event, list(self.childs.values())
+            )
+        if clients == EventTarget.self_:
+            return self.event_handler.fire_event(event, [self])
+        if clients == EventTarget.parent:
+            if self.parent is None:
+                raise RuntimeError(
+                    'object {} does not have parent'.format(self)
+                )
+            return self.event_handler.fire_event(event, None)
+
+    def add_child(self, child: 'EventClient') -> None:
+        child.event_handler = self.event_handler
+        child.parent = self
+        self.event_handler.register_client(child)
+        self.childs.update({child.uuid: child})
+        if self.is_running():
+            child.setup()
+
+    def delete_child(
+        self,
+        child: ty.Union['EventClient', UUID],
+        finalize: bool = True
+    ) -> None:
+        child = child if isinstance(child, EventClient) else self.childs[child]
+        if finalize:
+            child.cleanup()
+        del self.childs[child.uuid]
+
+    @reapy.inside_reaper()
+    def is_running(self) -> bool:
+        """Whether event loop is running."""
+        if self.parent is None:
+            return False
+        return self.parent.is_running()
+
+    def setup(self) -> None:
+        """User-method to be called at first mainloop iteration.
+
+        Note
+        ----
+        super().setup() has to be used at the end of overloaded method.
+        """
+        # for child in self.childs.values():
+        #     child.setup()
+
+    def _loop_frame(self) -> None:
+        """Reapy internal method to be called on every frame."""
+
+    def run(self) -> None:
+        """User-method to be called every loop frame.
+
+        Note
+        ----
+        super().run() has to be used at the end of overloaded method.
+        """
+        for child in self.childs.values():
+            child.run()
+
+    def cleanup(self) -> None:
+        """User-method to be called at exit.
+
+        Note
+        ----
+        super().cleanup() has to be used at the end of overloaded method.
+        """
+        for child in self.childs.values():
+            child.cleanup()
+        self.event_handler.remove_client(self)
+
+
+class EventLoop(EventClient):
+    """
+    Pure realization of event loop.
+
+
+    Methods to be overridden
+    -----------------------
+    def should_exit(self) -> bool: ...  True if loop has to be killed
+    def _launch(self) -> None: ...       as setup
+    def _loop_frame(self) -> None: ...   the same as run()
+    def _kill(self) -> None: ...         as cleanup
+    """
+
+    def __init__(self) -> None:
+        super().__init__(EventHandler())
+        self._running = False
+
+    @property
+    def _state(self) -> ty.Dict[str, object]:
+        return {**super()._state, "_running": self._running}
+
+    @reapy.inside_reaper()
+    def is_running(self) -> bool:
+        return self._running
+
+    def mainloop(self) -> None:
+        """Launch mainloop."""
+        self.start()
+        if not reapy.is_inside_reaper():
+            atexit.register(self._at_exit)
+            while self.is_running:
+                with reapy.inside_reaper():
+                    self.event_handler.fire_queue()
+
+    @reapy.inside_reaper()
+    def start(self) -> None:
+        """Launch event loop.
+
+        Not the best way to make event loop.
+        Runs internally, but, theoretically, can be used.
+        """
+        self._running = True
+        self._launch()
+        reapy.at_exit(self._at_exit)
+        self.fire_event(EvStart(), EventTarget.everyone)
+        self._loop()
+
+    def _launch(self) -> None:
+        """Reapy internal method to be called on start."""
+
+    @reapy.inside_reaper()
+    def _loop(self) -> None:
+        if self.should_exit() is True is self.is_running():
+            self._kill()
+            return
+        try:
+            self._loop_frame()
+            self.fire_event(EvFrame(), EventTarget.everyone)
+            reapy.defer(self._loop)
+        except KeyboardInterrupt as e:
+            self._kill()
+            raise e
+
+    def should_exit(self) -> bool:
+        """Whether event loop should be killed."""
+        return False
+
+    def _loop_frame(self) -> None:
+        """Reapy internal method to be called on every frame."""
+        for child in self.childs.values():
+            child._loop_frame()
+
+    @reapy.inside_reaper()
+    def _at_exit(self) -> None:
+        self.fire_event(EvExit(), EventTarget.everyone)
+        if not self.is_running():
+            return
+        self._kill()
+
+    def _kill(self) -> None:
+        """Reapy internal method to be called for cleanup."""
