+++ reapy/core/gui/themes.py
@@ -1,0 +1,32 @@
+import reapy
+from reapy import reascript_api as RPR
+from reapy.core import ReapyObject
+import typing as ty
+
+class Theme(ReapyObject):
+    main_bg = 
+
+    @property
+    def main_bg(self):
+        return -2134917189
+
+    @property
+    def main_text(self):
+        return 2236958
+
+    @property
+    def main_text_shadow(self):
+        return 3355443
+
+    @property
+    def main_3d_hl(self):
+        return 3355443
+
+    @property
+    def main_3d_sh(self):
+        return self._main_3d_sh
+    
+    
+    
+    
+
