+++ reapy/core/reapy_object.pyi
@@ -36,11 +36,10 @@ class ReapyObject:
         ...
 
     @property
-    def _state(self) -> ty.Optional[ty.Dict[str, object]]:
+    def _state(self) -> ty.Dict[str, object]:
         ...
 
-    @_state.setter
-    def _state(self, state: ty.Optional[ty.Dict[str, object]]) -> None: ...
+    def _state_set(self, state: ty.Dict[str, object]) -> None: ...
 
     def _to_dict(self) -> ReapyObjectDict:
         ...
